+++ lib/editor.ts
@@ -1,29 +1,43 @@
 import { EditorInterface, EditorLocaleSettings, SharedEditorSDK } from './types'
 import { Channel } from './channel'
 import { MemoizedSignal } from './signal'
+import { ConnectMessage } from './types/api.types'
 
 export default function createEditor(
   channel: Channel,
-  editorInterface: EditorInterface
+  editorInterface: EditorInterface,
+  editorData: Exclude<ConnectMessage['editor'], undefined>
 ): SharedEditorSDK['editor'] {
-  const _localeSettingsSygnal = new MemoizedSignal(undefined)
-  const _showDisabledFieldsSygnal = new MemoizedSignal(undefined)
+  const localeSettingsSignal = new MemoizedSignal<[EditorLocaleSettings]>(editorData.localeSettings)
+  const showHiddenFieldsSignal = new MemoizedSignal<[boolean]>(editorData.showHiddenFields)
 
   channel.addHandler('localeSettingsChanged', (settings: EditorLocaleSettings) => {
-    _localeSettingsSygnal.dispatch(settings)
+    localeSettingsSignal.dispatch(settings)
   })
 
-  channel.addHandler('showDisabledFieldsChanged', (showDisabledFields: boolean) => {
-    _showDisabledFieldsSygnal.dispatch(showDisabledFields)
+  channel.addHandler('showHiddenFieldsChanged', (showHiddenFields: boolean) => {
+    showHiddenFieldsSignal.dispatch(showHiddenFields)
   })
 
   return {
     editorInterface,
-    onLocaleSettingsChanged: (handler: Function) => {
-      return _localeSettingsSygnal.attach(handler)
+    getLocaleSettings(): EditorLocaleSettings {
+      return localeSettingsSignal.getMemoizedArgs()[0]
     },
-    onShowDisabledFieldsChanged: (handler: Function) => {
-      return _showDisabledFieldsSygnal.attach(handler)
+    onLocaleSettingsChanged: (handler) => {
+      return localeSettingsSignal.attach(handler)
+    },
+    /**
+     * @deprecated
+     */
+    onShowDisabledFieldsChanged: (handler) => {
+      return showHiddenFieldsSignal.attach(handler)
+    },
+    getShowHiddenFields(): boolean {
+      return showHiddenFieldsSignal.getMemoizedArgs()[0]
+    },
+    onShowHiddenFieldsChanged: (handler) => {
+      return showHiddenFieldsSignal.attach(handler)
     },
   }
 }