+++ lib/types/field-locale.types.ts
@@ -1,21 +1,18 @@
 import { ContentTypeFieldValidation } from './entities'
-import type { Items, SerializedJSONValue } from './utils'
+import type { SerializedJSONValue, FieldType, Items, FieldLinkType } from './utils'
 import { ValidationError } from './validation-error'
 
-export interface FieldAPI {
+interface FieldAPIBase {
   /** The ID of a field is defined in an entry's content type. */
   id: string
+  /** The name of the field */
+  name: string
   /** The current locale of a field the app is attached to. */
   locale: string
-  /** Holds the type of the field the app is attached to. */
-  type: string
   /** Indicates if a value for this field is required */
   required: boolean
   /** A list of validations for this field that are defined in the content type. */
   validations: ContentTypeFieldValidation[]
-  /** Defines the shape of array items */
-  items?: Items
-
   /** Gets the current value of the field and locale. */
   getValue: () => any
   /** Sets the value for the field and locale.  */
@@ -29,12 +26,61 @@ export interface FieldAPI {
    *  the returned function can be called to remove the handler function
    **/
   onValueChanged: (callback: (value: any) => void) => () => void
-  /** Calls the callback when the disabled status of the field changes.
-   *  the returned function can be called to remove the handler function
-   **/
+
+  /**
+   * Returns whether the current field is disabled
+   *
+   * The disabled state can change. To always work with the latest state, use `onIsDisabledChanged`.
+   */
+  getIsDisabled(): boolean
+
+  /**
+   * Subscribes to changes to this field's disabled state
+   *
+   * @param callback Function that is called every time this field's disabled state changes. Called immidiately with the current state.
+   * @returns Function to unsubscribe. `callback` won't be called anymore.
+   */
   onIsDisabledChanged: (callback: (isDisabled: boolean) => void) => () => void
-  /** Calls the callback immediately with the current validation errors and whenever the field is re-validated.
-   *  the returned function can be called to remove the handler function
-   **/
+
+  /**
+   * Returns the field's validation errors
+   *
+   * The schema errors can change. To always work with the latest errors, use `onSchemaErrorsChanged`.
+   */
+  getSchemaErrors(): ValidationError[]
+
+  /**
+   * Subscribes to schema errors
+   *
+   * @param callback Function that is called every time the schema errors change. Called immediately with the current errors.
+   * @returns Function to unsubscribe. `callback` won't be called anymore.
+   */
   onSchemaErrorsChanged: (callback: (errors: ValidationError[]) => void) => () => void
 }
+
+interface BasicFieldAPI extends FieldAPIBase {
+  /** Holds the type of the field. */
+  type: Exclude<FieldType, 'Array' | 'Link'>
+}
+
+interface ArrayFieldAPI extends FieldAPIBase {
+  /** Holds the type of the field. */
+  type: 'Array'
+  /** Defines the shape of array items */
+  items: Items
+}
+
+interface LinkFieldAPI extends FieldAPIBase {
+  /** Holds the type of the field. */
+  type: 'Link'
+  /** Type of linked resource */
+  linkType: FieldLinkType
+}
+
+export interface ExhaustiveFieldAPI extends FieldAPIBase {
+  type: FieldType
+  items?: Items
+  linkType?: FieldLinkType
+}
+
+export type FieldAPI = BasicFieldAPI | ArrayFieldAPI | LinkFieldAPI